SMS Gateway Plugin Update and Fix

+++ platform/plugins/sms/database/migrations/2022_12_12_153516_add_column_to_sms_table.php
@@ -14,8 +14,13 @@ class AddColumnToSmsTable extends Migration
     public function up()
     {
         Schema::table('sms', function (Blueprint $table) {
-            $table->text('template')->nullable();
-            $table->string('template_id')->nullable();
+            if (! Schema::hasColumn('sms', 'template')) {
+                $table->text('template')->nullable();
+            }
+
+            if (! Schema::hasColumn('sms', 'template_id')) {
+                $table->string('template_id')->nullable();
+            }
         });
     }
 
@@ -27,8 +32,13 @@ class AddColumnToSmsTable extends Migration
     public function down()
     {
         Schema::table('sms', function (Blueprint $table) {
-            $table->dropColumn('template');
-            $table->dropColumn('template_id');
+            if (Schema::hasColumn('sms', 'template')) {
+                $table->dropColumn('template');
+            }
+
+            if (Schema::hasColumn('sms', 'template_id')) {
+                $table->dropColumn('template_id');
+            }
         });
     }
 }
